requirements.txt
Flask
pybase64
jsonpickle
Pillow
numpy
opencv-python
#opencv-python-headless
fastapi
tflite
tensorflow
tensorflow_hub
#tflite-runtime
matplotlib
requests